+++ src/FakeXrmEasy.Core/XrmFakedContext.Crud.cs
@@ -174,8 +174,45 @@ namespace FakeXrmEasy
                     _service.Execute(request);
                 }
             }
+            
+            DeleteAssociatedFilesAfterUpdate(e);
         }
 
+        /// <summary>
+        /// Deletes any associated files to an entity that has their file attributes as null
+        /// </summary>
+        /// <param name="e"></param>
+        private void DeleteAssociatedFilesAfterUpdate(Entity e)
+        {
+            var associatedFiles = FileDb.GetFilesForTarget(e.ToEntityReference());
+            
+            foreach (var updatedAttribute in e.Attributes.Keys)
+            {
+                if (e[updatedAttribute] == null)
+                {
+                    var associatedFile = associatedFiles.FirstOrDefault(f => f.AttributeName.Equals(updatedAttribute));
+                    if (associatedFile != null)
+                    {
+                        FileDb.DeleteFile(associatedFile.Id);
+                    }
+                }
+            }
+        }
+        
+        /// <summary>
+        /// Deletes any associated files after a Delete message
+        /// </summary>
+        /// <param name="er">The entity reference that was deleted</param>
+        private void DeleteAssociatedFiles(EntityReference er)
+        {
+            var associatedFiles = FileDb.GetFilesForTarget(er);
+            
+            foreach (var associatedFile in associatedFiles)
+            {
+                FileDb.DeleteFile(associatedFile.Id);
+            }
+        }
+        
         /// <summary>
         /// Returns an entity record by logical name and primary key
         /// </summary>
@@ -315,6 +352,8 @@ namespace FakeXrmEasy
                 // Entity found => remove entity
                 var table = Db.GetTable(er.LogicalName);
                 table.Remove(er.Id);
+
+                DeleteAssociatedFiles(er);
             }
             else
             {